+++ lib/src/api/api_error_exception.dart
@@ -3,6 +3,7 @@ import 'package:dio/dio.dart';
 class ApiErrorException implements Exception {
   final String message;
   final bool isServerError;
+  final bool isSimpleError;
   final int? statusCode;
   final dynamic responseData;
   final DioException? dioError;
@@ -10,6 +11,7 @@ class ApiErrorException implements Exception {
   const ApiErrorException({
     required this.message,
     required this.isServerError,
+    required this.isSimpleError,
     this.statusCode,
     this.responseData,
     this.dioError,